Refactored of code in all adapter

+++ app/src/main/java/com/applocum/connecttomyhealth/ui/payment/models/FeaturePriceList.kt
@@ -4,6 +4,6 @@ import com.google.gson.annotations.SerializedName
 data class FeaturePriceList (
 
 	@SerializedName("label") val label : String,
-	@SerializedName("price") val price : Int,
+	@SerializedName("price") val price : Double,
 	@SerializedName("feature") val feature : String
 )
